requirements.txt
trino
pandas
numpy
scipy
python-dotenv
websockets